Implement loading of DB when starting flow

+++ db/version_manager.cc
@@ -9,7 +9,6 @@
 // libC++
 #include <algorithm>
 #include <cassert>
-
 #include <filesystem>
 
 namespace fs = std::filesystem;
@@ -31,7 +30,6 @@ VersionManager::VersionManager(
 void VersionManager::RemoveObsoleteVersion(uint64_t version_id) {
   std::scoped_lock lock(mutex_);
   auto it = versions_.find(version_id);
-
   if (it == versions_.end()) {
     return;
   }
@@ -48,7 +46,9 @@ void VersionManager::RemoveObsoleteVersion(uint64_t version_id) {
         // If ref count of a SST file = 0, it means that versions refer to it no
         // more. Time to say goodbye!
         fs::path file_path(sst_metadata[level][file_index]->filename);
-        fs::remove(file_path);
+        if (fs::exists(file_path) && fs::is_regular_file(file_path)) {
+          fs::remove(file_path);
+        }
       }
     }
   }
@@ -56,21 +56,81 @@ void VersionManager::RemoveObsoleteVersion(uint64_t version_id) {
   versions_.erase(it);
 }
 
-void VersionManager::CreateLatestVersion() {
-  if (!latest_version_) {
-    latest_version_ = std::make_unique<Version>(++next_version_id_, config_,
-                                                thread_pool_, this);
-    latest_version_->IncreaseRefCount();
-  }
-}
-
 void VersionManager::ApplyNewChanges(
     std::unique_ptr<VersionEdit> version_edit) {
   std::scoped_lock lock(mutex_);
 
   assert(version_edit);
-  assert(latest_version_);
 
+  if (!latest_version_) {
+    InitVersionWhenLoadingDb(std::move(version_edit));
+    return;
+  }
+
+  CreateNewVersion(std::move(version_edit));
+}
+
+void VersionManager::InitVersionWhenLoadingDb(
+    std::unique_ptr<VersionEdit> version_edit) {
+  uint64_t new_verions_id = ++next_version_id_;
+  auto new_version =
+      std::make_unique<Version>(new_verions_id, config_, thread_pool_, this);
+
+  std::vector<std::vector<std::shared_ptr<SSTMetadata>>>
+      &latest_version_sst_info = new_version->GetSSTMetadata();
+  std::vector<double> &latest_levels_score = new_version->GetLevelsScore();
+
+  const std::vector<std::vector<std::shared_ptr<SSTMetadata>>> &add_files =
+      version_edit->GetImmutableNewFiles();
+
+  for (int level = 0; level < add_files.size(); level++) {
+    for (const auto &sst_info : add_files[level]) {
+      // Increase refcount of SST metadata each time a new version is
+      // created
+      sst_info->ref_count++;
+
+      latest_version_sst_info[level].push_back(sst_info);
+    }
+
+    // Calcuate score ranking of each level
+    latest_levels_score[level] =
+        static_cast<double>(latest_version_sst_info[0].size()) /
+        static_cast<double>(config_->GetLvl0SSTCompactionTrigger());
+
+    if (level >= 1) {
+      // All SST files level >=1 must be sorted base on smallest key.
+      // Note : Because files levels >= 1 don't overlap with each other, so use
+      // smallest_key for condition to sort is enough
+      std::sort(latest_version_sst_info[level].begin(),
+                latest_version_sst_info[level].end(),
+                [](const auto &a, const auto &b) {
+                  return a->smallest_key < b->smallest_key;
+                });
+    }
+  }
+
+  latest_version_ = std::move(new_version);
+  // Each new version created has its refcount = 1
+  latest_version_->IncreaseRefCount();
+
+  // Remove obsolete SST files
+  thread_pool_->Enqueue(
+      [version_edit_ = std::move(version_edit), config_ = this->config_]() {
+        const std::set<std::pair<SSTId, int>> deleted_files =
+            version_edit_->GetImmutableDeletedFiles();
+        for (const auto &file : deleted_files) {
+          std::string filename =
+              config_->GetSavedDataPath() + std::to_string(file.first) + ".sst";
+          fs::path file_path(filename);
+          if (fs::exists(file_path) && fs::is_regular_file(file_path)) {
+            fs::remove(file_path);
+          }
+        }
+      });
+}
+
+void VersionManager::CreateNewVersion(
+    std::unique_ptr<VersionEdit> version_edit) {
   uint64_t new_verions_id = ++next_version_id_;
   auto new_version =
       std::make_unique<Version>(new_verions_id, config_, thread_pool_, this);
@@ -93,8 +153,8 @@ void VersionManager::ApplyNewChanges(
   // Apply all ssts info of previous version
   for (int level = 0; level < config_->GetSSTNumLvels(); level++) {
     for (const auto &sst_info : old_version_sst_info[level]) {
-      // Get score ranking from previous version(to know which level should be
-      // compacted)
+      // Get score ranking from previous version(to know which level should
+      // be compacted)
       latest_levels_score[level] = old_levels_score[level];
 
       // If file are in list of should be deleted file, skip
@@ -103,7 +163,8 @@ void VersionManager::ApplyNewChanges(
         continue;
       }
 
-      // Increase refcount of SST metadata each time a new version is created
+      // Increase refcount of SST metadata each time a new version is
+      // created
       sst_info->ref_count++;
 
       latest_version_sst_info[level].push_back(sst_info);
@@ -130,8 +191,8 @@ void VersionManager::ApplyNewChanges(
       continue;
     }
 
-    // With level >= 1. add new file and sort based on smallest key in ascending
-    // order
+    // With level >= 1. add new file and sort based on smallest key in
+    // ascending order
     const std::vector<std::shared_ptr<SSTMetadata>> &added_files_at_level =
         added_files[level];
     std::vector<std::shared_ptr<SSTMetadata>> new_sst_files;
@@ -147,6 +208,11 @@ void VersionManager::ApplyNewChanges(
                });
     // Assign back
     latest_version_sst_info[level] = std::move(new_sst_files);
+
+    // // Update level's score
+    // latest_levels_score[level] =
+    //     static_cast<double>(latest_version_sst_info[level].size()) /
+    //     static_cast<double>(config_->GetLvl0SSTCompactionTrigger());
   }
 
   // Update level 0' score
